requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
redis
pydantic
python-multipart
gunicorn
sqlalchemy
psycopg2-binary
google-auth
google-api-python-client